Refaktorering: Mer kode ut av server.ts
