+++ server/src/config/meta.ts
@@ -1,0 +1,1 @@
+export const basePath = "/min-ia";
