+++ server/src/environment.ts
@@ -1,0 +1,3 @@
+const isProduction = () => {
+  return process.env.NODE_ENV === "production";
+};
